Rewrite login queue.
The new implementation uses a bucket-based
system. Each bucket has a default capacity
of 100 players. The first two buckets in
queue are sent real-time updates on their
position, while the rest only have their
position updated with a precision of 100
when the first bucket is emptied and all
buckets are moved forward.

Login queue capacity and bucket size
are configurable.